fix(transcript): transcript now uses long

+++ src/lib-services/tagion/services/replicator.d
@@ -41,7 +41,7 @@ struct ReplicatorService {
         RecorderChainStorage storage = new RecorderChainFileStorage(opts.folder_path, net);
         RecorderChain recorder_chain = new RecorderChain(storage);
 
-        void receiveRecorder(SendRecorder, immutable(RecordFactory.Recorder) recorder, Fingerprint bullseye, immutable(int) epoch_number) {
+        void receiveRecorder(SendRecorder, immutable(RecordFactory.Recorder) recorder, Fingerprint bullseye, immutable(long) epoch_number) {
             auto last_block = recorder_chain.getLastBlock;
             auto block = new RecorderChainBlock(
                     recorder.toDoc,
